Use a different Travis script plus more
We should test with Cabal, not stack, because it allows to test in more
adverse environment, test with GHC versions that are not yet in snapshorts,
and generally improve compatibility.

This script tests with Cabal 1.24, as it's compatible with all GHC versions
of interest and has a better dependency resolution algorithm that allows to
avoid some problems with builds.

I also would like to impose 100% Haddock coverage here.